requirements.txt
django
pillow
django-debug-toolbar
django-extensions
django-bootstrap4
ipython
django-allauth
djangorestframework
django-cors-headers-3.7.0
